notification_service/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
PyJWT
pika